Decimal update and class fixing

+++ src/attitude-utils.cpp
@@ -39,7 +39,7 @@ Quaternion::Quaternion(const Vec3 &input) {
 }
 
 /// Create a quaternion which represents a rotation of theta around the axis input
-Quaternion::Quaternion(const Vec3 &input, float theta) {
+Quaternion::Quaternion(const Vec3 &input, decimal theta) {
     real = cos(theta/2);
     // the compiler will optimize it. Right?
     i = input.x * sin(theta/2);
@@ -54,7 +54,7 @@ Vec3 Quaternion::Rotate(const Vec3 &input) const {
 }
 
 /// How many radians the rotation represented by this quaternion has.
-float Quaternion::Angle() const {
+decimal Quaternion::Angle() const {
     if (real <= -1) {
         return 0; // 180*2=360=0
     }
@@ -63,7 +63,7 @@ float Quaternion::Angle() const {
 }
 
 /// Change the amount of rotation in a quaternion, keeping that rotation around the same axis.
-void Quaternion::SetAngle(float newAngle) {
+void Quaternion::SetAngle(decimal newAngle) {
     real = cos(newAngle/2);
     SetVector(Vector().Normalize() * sin(newAngle/2));
 }
@@ -76,18 +76,18 @@ EulerAngles Quaternion::ToSpherical() const {
     // and 2, we store the conjugate of the quaternion (double check why?), which means we need to
     // invert the final de and roll terms, as well as negate all the terms involving a mix between
     // the real and imaginary parts.
-    float ra = atan2(2*(-real*k+i*j), 1-2*(j*j+k*k));
+    decimal ra = atan2(2*(-real*k+i*j), 1-2*(j*j+k*k));
     if (ra < 0)
         ra += 2*M_PI;
-    float de = -asin(2*(-real*j-i*k)); // allow de to be positive or negaive, as is convention
-    float roll = -atan2(2*(-real*i+j*k), 1-2*(i*i+j*j));
+    decimal de = -asin(2*(-real*j-i*k)); // allow de to be positive or negaive, as is convention
+    decimal roll = -atan2(2*(-real*i+j*k), 1-2*(i*i+j*j));
     if (roll < 0)
         roll += 2*M_PI;
 
     return EulerAngles(ra, de, roll);
 }
 
-Quaternion SphericalToQuaternion(float ra, float dec, float roll) {
+Quaternion SphericalToQuaternion(decimal ra, decimal dec, decimal roll) {
     assert(roll >= 0.0 && roll <= 2*M_PI);
     assert(ra >= 0 && ra <= 2*M_PI);
     assert(dec >= -M_PI && dec <= M_PI);
@@ -107,7 +107,7 @@ Quaternion SphericalToQuaternion(float ra, float dec, float roll) {
 }
 
 /// Whether the quaternion is a unit quaternion. All quaternions representing rotations should be units.
-bool Quaternion::IsUnit(float tolerance) const {
+bool Quaternion::IsUnit(decimal tolerance) const {
     return abs(i*i+j*j+k*k+real*real - 1) < tolerance;
 }
 
@@ -123,7 +123,7 @@ Quaternion Quaternion::Canonicalize() const {
 }
 
 /// Convert from right ascension & declination to a 3d point on the unit sphere.
-Vec3 SphericalToSpatial(float ra, float de) {
+Vec3 SphericalToSpatial(decimal ra, decimal de) {
     return {
         cos(ra)*cos(de),
         sin(ra)*cos(de),
@@ -132,67 +132,67 @@ Vec3 SphericalToSpatial(float ra, float de) {
 }
 
 /// Convert from a 3d point on the unit sphere to right ascension & declination.
-void SpatialToSpherical(const Vec3 &vec, float *ra, float *de) {
+void SpatialToSpherical(const Vec3 &vec, decimal *ra, decimal *de) {
     *ra = atan2(vec.y, vec.x);
     if (*ra < 0)
         *ra += M_PI*2;
     *de = asin(vec.z);
 }
 
-float RadToDeg(float rad) {
+decimal RadToDeg(decimal rad) {
     return rad*180.0/M_PI;
 }
 
-float DegToRad(float deg) {
+decimal DegToRad(decimal deg) {
     return deg/180.0*M_PI;
 }
 
-float RadToArcSec(float rad) {
+decimal RadToArcSec(decimal rad) {
     return RadToDeg(rad) * 3600.0;
 }
 
-float ArcSecToRad(float arcSec) {
+decimal ArcSecToRad(decimal arcSec) {
     return DegToRad(arcSec / 3600.0);
 }
 
 /// The square of the magnitude
-float Vec3::MagnitudeSq() const {
+decimal Vec3::MagnitudeSq() const {
     return x*x+y*y+z*z;
 }
 
 /// The square of the magnitude
-float Vec2::MagnitudeSq() const {
+decimal Vec2::MagnitudeSq() const {
     return x*x+y*y;
 }
 
-float Vec3::Magnitude() const {
+decimal Vec3::Magnitude() const {
     return sqrt(MagnitudeSq());
 }
 
-float Vec2::Magnitude() const {
+decimal Vec2::Magnitude() const {
     return sqrt(MagnitudeSq());
 }
 
 /// Create a vector pointing in the same direction with magnitude 1
 Vec3 Vec3::Normalize() const {
-    float mag = Magnitude();
+    decimal mag = Magnitude();
     return {
         x/mag, y/mag, z/mag,
     };
 }
 
 /// Dot product
-float Vec3::operator*(const Vec3 &other) const {
+decimal Vec3::operator*(const Vec3 &other) const {
     return x*other.x + y*other.y + z*other.z;
 }
 
 /// Dot product
-Vec2 Vec2::operator*(const float &other) const {
+Vec2 Vec2::operator*(const decimal &other) const {
     return { x*other, y*other };
 }
 
 /// Vector-Scalar multiplication
-Vec3 Vec3::operator*(const float &other) const {
+Vec3 Vec3::operator*(const decimal &other) const {
     return { x*other, y*other, z*other };
 }
 
@@ -239,7 +239,7 @@ Vec3 Vec3::operator*(const Mat3 &other) const {
 }
 
 /// Access the i,j-th element of the matrix
-float Mat3::At(int i, int j) const {
+decimal Mat3::At(int i, int j) const {
     return x[3*i+j];
 }
 
@@ -283,7 +283,7 @@ Vec3 Mat3::operator*(const Vec3 &vec) const {
 }
 
 /// Matrix-Scalar multiplication
-Mat3 Mat3::operator*(const float &s) const {
+Mat3 Mat3::operator*(const decimal &s) const {
     return {
         s*At(0,0), s*At(0,1), s*At(0,2),
         s*At(1,0), s*At(1,1), s*At(1,2),
@@ -301,19 +301,19 @@ Mat3 Mat3::Transpose() const {
 }
 
 /// Trace of a matrix
-float Mat3::Trace() const {
+decimal Mat3::Trace() const {
     return At(0,0) + At(1,1) + At(2,2);
 }
 
 /// Determinant of a matrix
-float Mat3::Det() const {
+decimal Mat3::Det() const {
     return (At(0,0) * (At(1,1)*At(2,2) - At(2,1)*At(1,2))) - (At(0,1) * (At(1,0)*At(2,2) - At(2,0)*At(1,2))) + (At(0,2) * (At(1,0)*At(2,1) - At(2,0)*At(1,1)));
 }
 
 /// Inverse of a matrix
 Mat3 Mat3::Inverse() const {
     // https://ardoris.wordpress.com/2008/07/18/general-formula-for-the-inverse-of-a-3x3-matrix/
-    float scalar = 1 / Det();
+    decimal scalar = 1 / Det();
 
     Mat3 res = {
         At(1,1)*At(2,2) - At(1,2)*At(2,1), At(0,2)*At(2,1) - At(0,1)*At(2,2), At(0,1)*At(1,2) - At(0,2)*At(1,1),
@@ -355,7 +355,7 @@ Quaternion DCMToQuaternion(const Mat3 &dcm) {
     Vec3 newXAxis = dcm.Column(0); // this is where oldXAxis is mapped to
     assert(abs(newXAxis.Magnitude()-1) < 0.001);
     Vec3 xAlignAxis = oldXAxis.CrossProduct(newXAxis).Normalize();
-    float xAlignAngle = AngleUnit(oldXAxis, newXAxis);
+    decimal xAlignAngle = AngleUnit(oldXAxis, newXAxis);
     Quaternion xAlign(xAlignAxis, xAlignAngle);
 
     // Make a quaternion that will rotate the Y-axis into place
@@ -426,12 +426,12 @@ EulerAngles Attitude::ToSpherical() const {
 
 /// The length that a Vec3 will take up when serialized
 long SerializeLengthVec3() {
-    return sizeof(float)*3;
+    return sizeof(decimal)*3;
 }
 
 /// Serialize a Vec3 to buffer. Takes up space according to SerializeLengthVec3
 void SerializeVec3(const Vec3 &vec, unsigned char *buffer) {
-    float *fBuffer = (float *)buffer;
+    decimal *fBuffer = (decimal *)buffer;
     *fBuffer++ = vec.x;
     *fBuffer++ = vec.y;
     *fBuffer = vec.z;
@@ -439,7 +439,7 @@ void SerializeVec3(const Vec3 &vec, unsigned char *buffer) {
 
 Vec3 DeserializeVec3(const unsigned char *buffer) {
     Vec3 result;
-    const float *fBuffer = (float *)buffer;
+    const decimal *fBuffer = (decimal *)buffer;
     result.x = *fBuffer++;
     result.y = *fBuffer++;
     result.z = *fBuffer;
@@ -447,7 +447,7 @@ Vec3 DeserializeVec3(const unsigned char *buffer) {
 }
 
 /// Calculate the inner angle, in radians, between two vectors.
-float Angle(const Vec3 &vec1, const Vec3 &vec2) {
+decimal Angle(const Vec3 &vec1, const Vec3 &vec2) {
     return AngleUnit(vec1.Normalize(), vec2.Normalize());
 }
 
@@ -456,14 +456,14 @@ float Angle(const Vec3 &vec1, const Vec3 &vec2) {
  * Slightly faster than Angle()
  * @warn If the vectors are not already unit vectors, will return the wrong result!
  */
-float AngleUnit(const Vec3 &vec1, const Vec3 &vec2) {
-    float dot = vec1*vec2;
+decimal AngleUnit(const Vec3 &vec1, const Vec3 &vec2) {
+    decimal dot = vec1*vec2;
     // TODO: we shouldn't need this nonsense, right? how come acos sometimes gives nan?
     return dot >= 1 ? 0 : dot <= -1 ? M_PI-0.0000001 : acos(dot);
 }
 
 /// The distance between two vectors, according to the usual distance formula.
-float Distance(const Vec2 &v1, const Vec2 &v2) {
+decimal Distance(const Vec2 &v1, const Vec2 &v2) {
     return sqrt(pow(v1.x-v2.x, 2) + pow(v1.y-v2.y, 2));
 }
